+++ src/envConfig.ts
@@ -396,6 +396,11 @@ export const asEnvConfig = asObject({
       apiKey: asOptional(asString, '')
     }).withRest
   ),
+  XGRAM_INIT: asCorePluginInit(
+    asObject({
+      apiKey: asOptional(asString, '')
+    }).withRest
+  ),
   UNIZEN_INIT: asCorePluginInit(
     asObject({
       apiKey: asOptional(asString, '')